+++ includes/Admin/DiagnosticsPage.php
@@ -98,7 +98,9 @@ class DiagnosticsPage {
                 $dupes    = array_filter( $counts, static fn( $c ) => $c > 1 );
                 $render_ok = empty( $dupes );
 
-                $dup_plugins        = Install::duplicates();
+                $install_status     = Install::detectDuplicates();
+                $dup_plugins        = $install_status['duplicates'];
+                $canonical_slug     = $install_status['canonical'];
                 $last_consolidation = Install::log();
                 /* @psalm-suppress UnresolvableInclude */
                 require FBM_PATH . 'templates/admin/diagnostics.php';